Refactors favorites index to use facade and updates authorize user model method to include locations
